+++ src/components/components/HighlightText.tsx
@@ -1,0 +1,18 @@
+export const HighlightText = ({ text, className = '' }) => {
+  const parts = text.split(/(\*\*.*?\*\*)/);
+
+  return (
+    <span className={className}>
+      {parts.map((part, index) => {
+        if (part.startsWith('**') && part.endsWith('**')) {
+          return (
+            <span key={index} className="text-primary">
+              {part.slice(2, -2)}
+            </span>
+          );
+        }
+        return <span key={index}>{part}</span>;
+      })}
+    </span>
+  );
+};
