+++ src/components/coupons/HeroSection.tsx
@@ -6,6 +6,7 @@ import Countdown from '../components/Countdown';
 import { currencyService } from '@/services/currency.service';
 import { useEffect, useState } from 'react';
 import { CouponPageText } from '@/assets/types/couponsPage';
+import { HighlightText } from '../components/HighlightText';
 
 interface HeroSectionForHomeProps {
   textContent: CouponPageText['HeroSection'];
@@ -70,25 +71,6 @@ export default function HeroSection({
       return <div className="bg-gray-200 h-4 w-16 animate-pulse rounded"></div>;
     }
     return typeof text === 'string' ? text.replace(/{{discount}}/g, percentOff) : text;
-  };
-
-  const HighlightText = ({ text, className = '' }) => {
-    const parts = text.split(/(\*\*.*?\*\*)/);
-
-    return (
-      <span className={className}>
-        {parts.map((part, index) => {
-          if (part.startsWith('**') && part.endsWith('**')) {
-            return (
-              <span key={index} className="text-primary">
-                {part.slice(2, -2)}
-              </span>
-            );
-          }
-          return <span key={index}>{part}</span>;
-        })}
-      </span>
-    );
   };
 
   return (
