+++ src/components/meet/HeroSection.tsx
@@ -1,27 +1,7 @@
 import { getImage } from '@/lib/getImage';
 import Image from 'next/image';
-import styles from '@/components/privacy/HeroSection.module.scss';
-import HeroSafeArea from '../HeroSafeArea';
 import Link from 'next/link';
-
-const HighlightText = ({ text, className = '' }) => {
-  const parts = text.split(/(\*\*.*?\*\*)/);
-
-  return (
-    <span className={className}>
-      {parts.map((part, index) => {
-        if (part.startsWith('**') && part.endsWith('**')) {
-          return (
-            <span key={index} className="text-primary">
-              {part.slice(2, -2)}
-            </span>
-          );
-        }
-        return <span key={index}>{part}</span>;
-      })}
-    </span>
-  );
-};
+import { HighlightText } from '../components/HighlightText';
 
 const HeroSection = ({ textContent }) => (
   <section
