+++ src/components/nas/HeroSection.tsx
@@ -3,25 +3,7 @@ import Image from 'next/image';
 import styles from '@/components/privacy/HeroSection.module.scss';
 import Link from 'next/link';
 import { Check } from '@phosphor-icons/react';
-
-const HighlightText = ({ text, className = '' }) => {
-  const parts = text.split(/(\*\*.*?\*\*)/);
-
-  return (
-    <span className={className}>
-      {parts.map((part, index) => {
-        if (part.startsWith('**') && part.endsWith('**')) {
-          return (
-            <span key={index} className="text-primary">
-              {part.slice(2, -2)}
-            </span>
-          );
-        }
-        return <span key={index}>{part}</span>;
-      })}
-    </span>
-  );
-};
+import { HighlightText } from '../components/HighlightText';
 
 const HeroSection = ({ textContent }) => (
   <section
